Added world name to the constants file.
Primarily worked inside of the dataManager. this is what is used to do any of the data related functions (get, put, etc.) So far I've created a function to create a new directory if one doesn't exist for the module and another function to create a file as long as you supply the data, and stores it as a JSON.

I've also created a function to clean the backup folder and gets backup files. The clean function only performs a clean for the system you are currently using (doesn't delete backups from other systems) and it leaves the last 4 files and deletes the rest.

Get backup file doesn't take any arguments and is intended to only be ran when foundry first loads to populate the module with data.