IMPORTANT added types from Foundry Jsdoc
for this to work, sys liinks need to be made from the foundry application directory of the common and client directories
and they should be linked to a folder called external in the root of the module. The jsconfig.json file will include
those files and their exports for the lsp for completion.

Implementation of a View Manager

TODO:
* Implement state management on the class, to track things like whether the application is opened, rendered, etc.
* Application components should be able to be passed into this class for rendering, opening, closing, and state management

Added an EncounterModel for managing the data layer of encounters within Foundry

Re desigining the scope of the knowledge Recalled class

This class should be responsible for for initializing the application, managing hooks and their states, but things that
occur on any of the data models should occur their rather in this location. Anything pertaining to hooks likely need
there own HookManagement class which should initialize some sort of event listener and maintain a singleton state.

Experimenting here mainly trying to get this to render, but isn't quite ready for implementation

UI override controller

experimenting here

formatting and added depenency

+Refactor - Core - Centralized all of the hooks into a function called registerHooks()

+ Core - ViewManager Class for managing the application

Implemented registeringApplications, registeringProtectedApplications, open and close applications, and a close all
applications method. Implemented a protection state property that restricts the standard closeAll and close method
from closing those applications. Intended for HUD and protected UI elements. We can implement a method for explicitly
closing protected applications. This can be toggled in via a button should people wish to hide the HUD. Implement
A pin feature where we can also toggle protection on elements that don't typically have it.

* Formatting - no changes

* formatting

+ Core - Implemented a method for getting Active encounters

Also listed methods that need to be depricated and either replaced and reworked or need to be removed once there
functionality is no longer needed and ensured to not be needed.

+ Refactor - Moved rendering responsibility to The ViewManager

This should make opening and closing the application easier. Need to get working with injecting the icon.

removed console log

deleted appService

initialliazation should make that never undefined

Function that autoNegotiates whether to open or close based on current state

Method to manually alter the opened value state, that way we can change state even if it occurs outside of our module

added auto negotiate to ui button

added hook for when application closes

Created utility file for some functional commands of getting data

separated getting encoutners from the class, began the populate GM journal method

Removed all hooks to the hooks file