IMPORTANT added types from Foundry Jsdoc
for this to work, sys liinks need to be made from the foundry application directory of the common and client directories
and they should be linked to a folder called external in the root of the module. The jsconfig.json file will include
those files and their exports for the lsp for completion.