(fix) Resolved View Manager and LSP iss
(fix) Missing package, eslint config

Issue with the eslint config not being a valid arguement, I also was
missing a dependency that I have since adding. this was causing issues
with the LSP server, but it now appears to be resolved.

(fix) resolved viewmanager issues

There was a typo in the class name of the GmJournalView object, and I
had to rework how this was working since I moved to embedding the API
inside of the module path provided by foundry.