(feat) improved logging
Logging methods have been improved! Now each log not only
provides better information, including stack trace for errors
it will also provide a running sessionLog on each log object
to see all the other logs the module has generated. This is refreshed
when Foundry refreshes and does not persist. Contemplating logging to
a file for persistence if it's valuable in the future.